Add Truck Cleanup on Home page: goes through each entry and sees if any of them share the same truck id. If they do, duplicates the result such that each entry has it's own unique truck id. Up version 3.2.4 Fix isFlowEntry

+++ server/play/app/models/Truck.java
@@ -5,6 +5,7 @@ package models;
 
 import java.util.*;
 import java.lang.Long;
+
 import org.apache.commons.lang3.StringUtils;
 
 import javax.persistence.*;
@@ -13,13 +14,14 @@ import play.db.ebean.*;
 import play.data.validation.*;
 import play.db.ebean.Transactional;
 import play.data.format.*;
+
 import com.avaje.ebean.PagedList;
 
 import play.Logger;
 
 /**
  * User entity managed by Ebean
- *
+ * <p>
  * NOTE: This whole class is an obsolete idea.
  * What has replaced this, is a unique note data value with an associated unique picture. No need
  * for an entire table.
@@ -119,9 +121,27 @@ public class Truck extends com.avaje.ebean.Model {
                 Logger.info("Cannot delete truck, has entries: " + truck.toString());
                 continue;
             }
+            if (WorkOrder.countWorkOrdersForTruck(truck.id) > 0) {
+                Logger.info("Cannot delete truck, used in work orders: " + truck.toString());
+                continue;
+            }
             Logger.info("Deleting truck: " + truck.toString());
             truck.delete();
             count++;
+        }
+        Logger.info("Deleted " + count + " trucks");
+        return count;
+    }
+
+    public static int cleanup2() {
+        int count = 0;
+        List<Truck> list = find.all();
+        Logger.info("Found " + list.size() + " trucks");
+        for (Truck truck : list) {
+            if (truck.countEntries() == 0 && WorkOrder.countWorkOrdersForTruck(truck.id) == 0) {
+                truck.delete();
+                count++;
+            }
         }
         Logger.info("Deleted " + count + " trucks");
         return count;
@@ -138,10 +158,10 @@ public class Truck extends com.avaje.ebean.Model {
     public static Truck add(long project_id, long company_id, long truck_id, String truck_number, String license_plate, int tech_id) {
         long company_name_id = 0;
         if (company_id > 0) {
-           Company company = Company.get(company_id);
-           if (company != null) {
-               company_name_id = CompanyName.save(company.name);
-           }
+            Company company = Company.get(company_id);
+            if (company != null) {
+                company_name_id = CompanyName.save(company.name);
+            }
         }
         Truck truck = null;
         List<Truck> list;
@@ -161,65 +181,65 @@ public class Truck extends com.avaje.ebean.Model {
                 truck = null;
             }
         }
-        if (truck == null) {
-            if (!StringUtils.isEmpty(truck_number)) {
-                if (project_id > 0 && company_name_id > 0) {
-                    list = find.where()
-                            .eq("project_id", project_id)
-                            .eq("company_name_id", company_name_id)
-                            .eq("truck_number", truck_number)
-                            .findList();
-                } else if (project_id > 0) {
-                    list = find.where()
-                            .eq("project_id", project_id)
-                            .eq("truck_number", truck_number)
-                            .findList();
-                } else if (company_name_id > 0) {
-                    list = find.where()
-                            .eq("company_name_id", company_name_id)
-                            .eq("truck_number", truck_number)
-                            .findList();
-                } else {
-                    list = find.where()
-                            .eq("truck_number", truck_number)
-                            .findList();
-                }
-            } else {
-                if (project_id > 0 && company_name_id > 0) {
-                    list = find.where()
-                            .eq("project_id", project_id)
-                            .eq("company_name_id", company_name_id)
-                            .eq("license_plate", license_plate)
-                            .findList();
-                } else if (project_id > 0) {
-                    list = find.where()
-                            .eq("project_id", project_id)
-                            .eq("license_plate", license_plate)
-                            .findList();
-                } else if (company_name_id > 0) {
-                    list = find.where()
-                            .eq("company_name_id", company_name_id)
-                            .eq("license_plate", license_plate)
-                            .findList();
-                } else {
-                    list = find.where()
-                            .eq("license_plate", license_plate)
-                            .findList();
-                }
-            }
-            if (list != null) {
-                if (list.size() > 1) {
-                    Logger.error("Found too many trucks with "
-                            + truck_number + ", " + license_plate + ", project_id=" + project_id + ", company_name_id=" + company_name_id);
-                    for (Truck t : list) {
-                        Logger.error(t.toString());
-                    }
-                    truck = list.get(0);
-                } else if (list.size() == 1) {
-                    truck = list.get(0);
-                }
-            }
-        }
+//        if (truck == null) {
+//            if (!StringUtils.isEmpty(truck_number)) {
+//                if (project_id > 0 && company_name_id > 0) {
+//                    list = find.where()
+//                            .eq("project_id", project_id)
+//                            .eq("company_name_id", company_name_id)
+//                            .eq("truck_number", truck_number)
+//                            .findList();
+//                } else if (project_id > 0) {
+//                    list = find.where()
+//                            .eq("project_id", project_id)
+//                            .eq("truck_number", truck_number)
+//                            .findList();
+//                } else if (company_name_id > 0) {
+//                    list = find.where()
+//                            .eq("company_name_id", company_name_id)
+//                            .eq("truck_number", truck_number)
+//                            .findList();
+//                } else {
+//                    list = find.where()
+//                            .eq("truck_number", truck_number)
+//                            .findList();
+//                }
+//            } else {
+//                if (project_id > 0 && company_name_id > 0) {
+//                    list = find.where()
+//                            .eq("project_id", project_id)
+//                            .eq("company_name_id", company_name_id)
+//                            .eq("license_plate", license_plate)
+//                            .findList();
+//                } else if (project_id > 0) {
+//                    list = find.where()
+//                            .eq("project_id", project_id)
+//                            .eq("license_plate", license_plate)
+//                            .findList();
+//                } else if (company_name_id > 0) {
+//                    list = find.where()
+//                            .eq("company_name_id", company_name_id)
+//                            .eq("license_plate", license_plate)
+//                            .findList();
+//                } else {
+//                    list = find.where()
+//                            .eq("license_plate", license_plate)
+//                            .findList();
+//                }
+//            }
+//            if (list != null) {
+//                if (list.size() > 1) {
+//                    Logger.error("Found too many trucks with "
+//                            + truck_number + ", " + license_plate + ", project_id=" + project_id + ", company_name_id=" + company_name_id);
+//                    for (Truck t : list) {
+//                        Logger.error(t.toString());
+//                    }
+//                    truck = list.get(0);
+//                } else if (list.size() == 1) {
+//                    truck = list.get(0);
+//                }
+//            }
+//        }
         if (truck == null) {
             truck = new Truck();
             truck.truck_number = truck_number;
@@ -392,18 +412,20 @@ public class Truck extends com.avaje.ebean.Model {
         StringBuilder sbuf = new StringBuilder();
         sbuf.append("ID ");
         sbuf.append(id);
-        sbuf.append(" ");
-        if (truck_number != null) {
+        sbuf.append(" NUMBER ");
+        if (truck_number != null && !truck_number.isEmpty()) {
             sbuf.append(truck_number);
         }
-        if (license_plate != null) {
+        if (license_plate != null && !license_plate.isEmpty()) {
             sbuf.append(" : ");
             sbuf.append(license_plate);
         }
         if (project_id > 0) {
-            sbuf.append(", PROJECT ");
-            sbuf.append(getRootProjectName());
-            sbuf.append(" - ");
+            if (getRootProjectName() != null) {
+                sbuf.append(", PROJECT ");
+                sbuf.append(getRootProjectName());
+                sbuf.append(" - ");
+            }
             sbuf.append(getSubProjectName());
         }
         if (company_name_id > 0) {
